Before compé last updated code

+++ Core/Inc/main.h
@@ -66,14 +66,14 @@ typedef struct SensorData_{
 	uint8_t limit1, limit2;
 
 	// Motor modes
-	uint8_t feedback_pitch_mode;
-	uint8_t feedback_mast_mode;
+	uint32_t feedback_pitch_mode;
+	uint32_t feedback_mast_mode;
 
-	uint8_t feedback_pitch_rops;
+	uint32_t feedback_pitch_rops;
 
 	// Motor faults
-	uint8_t pitch_fault_stall;
-	uint8_t mast_fault_stall;
+	uint32_t pitch_fault_stall;
+	uint32_t mast_fault_stall;
 
 	// Processed sensor data
 	float pitch_angle;
